+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Unreleased
 - Added support for overwriting default realtime data bar, quote, and
   trade types
 - Added support for working with custom v2 realtime data streaming URLs
+- Added `FillOrKill` and `ImmediateOrCancel` variants to
+  `api::v2::order::TimeInForce` type
 - Changed various `data::v2::stream::{Bar, Quote, Trade}` member from
   `u64` to `Num`
 - Switched to using new stream authentication request message format